+++ oakd_yolo.py
@@ -2,10 +2,7 @@
 
 import contextlib
 import json
-import os
 import time
-import requests
-from tqdm import tqdm
 from pathlib import Path
 from typing import Any, List, Optional, Tuple, Union
 
@@ -55,20 +52,20 @@ class OakdYolo(object):
         self.confidenceThreshold = metadata.get("confidence_threshold", {})
 
         print(metadata)
-        self.cam_fps = fps
+        self.fps = fps
         # parse labels
         nnMappings = config.get("mappings", {})
         self.labels = nnMappings.get("labels", {})
 
-        self.nnPath = Path(model_path)
+        self.nn_path = Path(model_path)
         # get model path
-        if not self.nnPath.exists():
+        if not self.nn_path.exists():
             print(
                 "No blob found at {}. Looking into DepthAI model zoo.".format(
-                    self.nnPath
+                    self.nn_path
                 )
             )
-            self.nnPath = str(
+            self.nn_path = str(
                 blobconverter.from_zoo(
                     model_path, shaves=6, zoo_type="depthai", use_cache=True
                 )
@@ -85,51 +82,17 @@ class OakdYolo(object):
         self.qRaw = self._device.getOutputQueue(name="raw")
         self.qDet = self._device.getOutputQueue(name="nn", maxSize=4, blocking=False)
         self.counter = 0
-        self.startTime = time.monotonic()
+        self.start_time = time.monotonic()
         self.frame_name = 0
         self.dir_name = ""
         self.path = ""
         self.num = 0
-        self.counter = 0
         self.raw_frame = None
+
 
     def close(self) -> None:
         """OAK-Dを閉じる。"""
         self._device.close()
-
-    def download_file(self, path: str, link: str) -> None:
-        """
-        指定されたリンクからファイルをダウンロードする関数。
-
-        Args:
-            path (str): ダウンロード先のファイルパス
-            link (str): ダウンロード元のリンク
-
-        Raises:
-            Exception: ダウンロード中にエラーが発生した場合
-        """
-        if os.path.exists(path):
-            return
-        # ディレクトリが存在しない場合は作成
-        dir_path = os.path.dirname(path)
-        if not os.path.exists(dir_path):
-            os.makedirs(dir_path)
-        try:
-            # ファイルをダウンロード
-            print(f"{path} doesn't exist. Download from {link}")
-            response = requests.get(link, stream=True)
-            total_size = int(response.headers.get("content-length", 0))
-            block_size = 1024
-            progress = tqdm(total=total_size, unit="iB", unit_scale=True)
-            with open(path, "wb") as f:
-                for data in response.iter_content(chunk_size=block_size):
-                    if data:
-                        f.write(data)
-                        progress.update(len(data))
-            progress.close()
-            print(f"{path} download finished.")
-        except Exception as e:
-            print(f"Download error")
 
     def set_camera_brightness(self, brightness: int) -> None:
         """カメラの明るさを設定する。
@@ -177,7 +140,7 @@ class OakdYolo(object):
         camRgb.setInterleaved(False)
         camRgb.setColorOrder(dai.ColorCameraProperties.ColorOrder.BGR)
         camRgb.setPreviewSize(1920, 1080)
-        camRgb.setFps(self.cam_fps)
+        camRgb.setFps(self.fps)
 
         xoutRaw = pipeline.create(dai.node.XLinkOut)
         xoutRaw.setStreamName("raw")
@@ -195,7 +158,7 @@ class OakdYolo(object):
         detectionNetwork.setAnchors(self.anchors)
         detectionNetwork.setAnchorMasks(self.anchorMasks)
         detectionNetwork.setIouThreshold(self.iouThreshold)
-        detectionNetwork.setBlobPath(self.nnPath)
+        detectionNetwork.setBlobPath(self.nn_path)
         detectionNetwork.setNumInferenceThreads(2)
         detectionNetwork.input.setBlocking(False)
 
@@ -337,7 +300,7 @@ class OakdYolo(object):
             cv2.putText(
                 frame,
                 "NN fps: {:.2f}".format(
-                    self.counter / (time.monotonic() - self.startTime)
+                    self.counter / (time.monotonic() - self.start_time)
                 ),
                 (2, frame.shape[0] - 4),
                 cv2.FONT_HERSHEY_TRIPLEX,
